delete unused SingleYield
