--- lib/mocha/single_yield.rb
@@ -1,13 +1,0 @@
-module Mocha
-  class SingleYield
-    attr_reader :parameters
-
-    def initialize(*parameters)
-      @parameters = parameters
-    end
-
-    def each
-      yield(@parameters)
-    end
-  end
-end
