+++ lib/mocha/yield_parameters.rb
@@ -1,5 +1,4 @@
 require 'mocha/no_yields'
-require 'mocha/single_yield'
 require 'mocha/multiple_yields'
 
 module Mocha
